+++ CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog for aim.models
 3.2.0 (unreleased)
 ------------------
 
+### Added
+
+ - Added a CloudTrail schema configured in `Resources/CloudTrail.yaml`.
+
+ - IS3BucketPolicy now has `principal` and `condition` fields.
+   `principal` can be either a Key-Value dictionary, where the key is either 'AWS', 'Service', etc.
+   and the value can be either a String or a List. It is an alternate to the `aws` field, which will
+   remain for setting simpler AWS-only principals.
+   The `condition` field is a Key-Value dictionary of Key-Value filters.
+
 ### Changed
 
  - `MonitorConfig/LogSets.yaml` has been renamed to `MonitorConfig/Logging.yaml`. CloudWatch